+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/productbatches/productBatchesList/ProductBatchesListScreen.kt
@@ -220,7 +220,6 @@ fun ProductBatchesListScreen(
         }
     }
 
-    // Filtro modal
     if (showFilter && state.filterData != null) {
         Filter(
             data = state.filterData!!,