+++ app/src/main/java/com/app/arcabyolimpo/domain/usecase/supplies/GetFilterBatchDataUseCase.kt
@@ -7,16 +7,28 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
 
-class GetFilterBatchDataUseCase @Inject constructor(
-    private val repository: SupplyRepository
-) {
-    operator fun invoke(): Flow<Result<FilterData>> = flow {
-        emit(Result.Loading)
-        try {
-            val filterData = repository.getFilterBatchData()
-            emit(Result.Success(filterData))
-        } catch (e: Exception) {
-            emit(Result.Error(e))
-        }
+/**
+ * Use case responsible for retrieving filter data for supply batches.
+ *
+ * @property repository The repository that communicates with the remote API.
+ * @constructor Creates an instance of [GetFilterBatchDataUseCase].
+ * @throws Exception If an error occurs during the process.
+ * @return A flow emitting the result of the operation.
+ * @see Result
+ */
+class GetFilterBatchDataUseCase
+    @Inject
+    constructor(
+        private val repository: SupplyRepository,
+    ) {
+        operator fun invoke(): Flow<Result<FilterData>> =
+            flow {
+                emit(Result.Loading)
+                try {
+                    val filterData = repository.getFilterBatchData()
+                    emit(Result.Success(filterData))
+                } catch (e: Exception) {
+                    emit(Result.Error(e))
+                }
+            }
     }
-}