+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [0.2.0] - 03/12/2025
+### Added
+- Delete disability
+
 ### Fixed
 - Disability's details' routes
 - Modify beneficiary (Active/inactive beneficiary)